refactor: split fitting.py into estimation and metrics modules
- Create estimation.py with parameter fitting functions (MLE, MSE, bootstrap)
- Create metrics.py with goodness-of-fit functions (KS, AD, AIC, BIC)
- Add create_truncated_dist helper to truncated.py
- Update fitting.py to re-export from new modules for backward compatibility

This improves code organization and separation of concerns while
maintaining full backward compatibility for existing imports.
